+++ lib/include/job_api_wireless_nxa.h
@@ -31,7 +31,9 @@
 #define JOB_API_WIRELESS_NXA_H
 
 void
-generate_hqp_snow5g_sse(const uint8_t *key, const uint8_t *iv, uint8_t *hqp);
+generate_hqp_snow5g_sse(const uint8_t *key, const uint8_t *iv, uint8_t *hqp, uint8_t *state);
+void
+snow5g_nca4_sse(IMB_JOB *job, uint8_t *state);
 
 static const uint8_t zero_low_4B_mask[16] = { 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
                                               0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00 };
@@ -172,7 +174,7 @@ submit_snow5g_nia4_job(IMB_JOB *job)
         const uint8_t *msg = job->src + job->hash_start_src_offset_in_bytes;
 
         /* Generate H, Q, P keys for SNOW5G NIA4 */
-        GENERATE_HQP_SNOW5G(job->u.SNOW5G_NIA4._key, job->u.SNOW5G_NIA4._iv, HQP);
+        GENERATE_HQP_SNOW5G(job->u.SNOW5G_NIA4._key, job->u.SNOW5G_NIA4._iv, HQP, NULL);
 
         /* Precompute hash keys from H */
         POLYVAL_PRE(H, &gdata_key);
@@ -202,6 +204,61 @@ submit_snow5g_nia4_job(IMB_JOB *job)
         clear_mem(&gdata_key, sizeof(struct gcm_key_data));
         clear_mem(digest, sizeof(digest));
 #endif
+        return job;
+}
+
+__forceinline IMB_JOB *
+submit_snow5g_nca4_job(IMB_JOB *job, IMB_CIPHER_DIRECTION cipher_dir)
+{
+        DECLARE_ALIGNED(uint8_t HQP[3 * 16], 64);
+        DECLARE_ALIGNED(uint8_t digest[16], 16) = { 0 };
+        DECLARE_ALIGNED(uint8_t state[10 * 16], 16);
+
+        struct gcm_key_data gdata_key;
+        uint8_t *H = HQP;
+        uint8_t *Q = &HQP[16];
+        uint8_t *P = &HQP[16 * 2];
+        uint8_t const *out = job->dst;
+        uint64_t msg_len_bytes = job->msg_len_to_cipher_in_bytes;
+        uint64_t msg_len_bits = msg_len_bytes * 8;
+        const uint8_t *msg = job->src + job->cipher_start_src_offset_in_bytes;
+
+        GENERATE_HQP_SNOW5G(job->enc_keys, job->iv, HQP, state);
+        POLYVAL_PRE(H, &gdata_key);
+
+        POLYVAL(&gdata_key, job->u.NCA.aad, job->u.NCA.aad_len_in_bytes, digest);
+
+        if (cipher_dir == IMB_DIR_ENCRYPT) {
+                SNOW5G_NCA4(job, state);
+                POLYVAL(&gdata_key, out, msg_len_bytes, digest);
+        } else {
+
+                POLYVAL(&gdata_key, msg, msg_len_bytes, digest);
+
+                SNOW5G_NCA4(job, state);
+        }
+
+        uint8_t lengths[16] = { 0 };
+        const uint64_t aad_len_bits = job->u.NCA.aad_len_in_bytes * 8;
+        memcpy(lengths, &msg_len_bits, 8);
+        memcpy(&lengths[8], &aad_len_bits, 8);
+
+        for (int i = 0; i < 16; i++)
+                digest[i] ^= lengths[i];
+
+        POLYVAL_16B(Q, digest);
+
+        for (int i = 0; i < 16; i++)
+                digest[i] ^= P[i];
+
+        memcpy(job->auth_tag_output, digest, job->auth_tag_output_len_in_bytes);
+
+#ifdef SAFE_DATA
+        clear_mem(state, sizeof(state));
+        clear_mem(HQP, sizeof(HQP));
+        clear_mem(&gdata_key, sizeof(struct gcm_key_data));
+#endif
+        job->status |= IMB_STATUS_COMPLETED;
         return job;
 }