+++ ReleaseNotes.md
@@ -30,6 +30,7 @@ Library
 - Added ZUC-NIA6 SSE type 1 implementation.
 - Added ZUC-NCA6 SSE type 1 and SSE type 3 implementations.
 - Added SNOW5G-NIA4 SSE type 1 implementation.
+- Added SNOW5G-NCA4 SSE type 1 implementation.
 
 Test Applications
 - Added `--offset` parameter to `imb-xvalid` application.